+++ src/pages/product-list-page/product-list-filter/product-list-filter.js
@@ -4,26 +4,16 @@ import Button from '@material-ui/core/Button';
 import { useDispatch, useSelector } from 'react-redux';
 import Grid from '@material-ui/core/Grid';
 import { useHistory, useLocation } from 'react-router';
-import { map } from 'lodash';
 import PriceFilter from './price-filter';
 import HotItemFilter from './hot-item-filter';
 import { useStyles } from './product-list-filter.styles';
-import {
-  getFiltredProducts,
-  setPatternsFilter,
-  setCategoryFilter,
-  setModelsFilter
-} from '../../../redux/products/products.actions';
+import { getFiltredProducts } from '../../../redux/products/products.actions';
 
-import {
-  MODEL_TEXT,
-  PATTERN_TEXT,
-  CATERGORY_TEXT,
-  CLEAR_FILTER_BUTTON_TEXT
-} from '../../../translations/product-list.translations';
+import { CLEAR_FILTER_BUTTON_TEXT } from '../../../translations/product-list.translations';
 import ProductsFiltersContainer from '../../../containers/products-filters-container';
 import { selectFilterData } from '../../../redux/selectors/multiple.selectors';
-import { countPerPage, page, sort, URL_QUERIES_NAME } from '../../../configs';
+import { countPerPage, sort } from '../../../configs';
+import useProductFilters from '../../../hooks/use-product-filters';
 
 const ProductListFilter = () => {
   const styles = useStyles();
@@ -31,97 +21,9 @@ const ProductListFilter = () => {
   const history = useHistory();
   const { search } = useLocation();
   const searchParams = new URLSearchParams(search);
-  const { filters, language, filterData } = useSelector(selectFilterData);
-  const { categoryFilter, patternsFilter, modelsFilter } = filters;
+  const { language } = useSelector(selectFilterData);
+  const filtersOptions = useProductFilters();
 
-  const handleFilterChange = ({ target }, queryName, categoriesList) => {
-    let query = searchParams.get(queryName);
-    if (categoriesList) {
-      const categoryId = categoriesList.filter(
-        (element) => element.name[language].value === target.name
-      )[0]._id;
-      if (query) {
-        if (!target.checked) {
-          query = query.replace(categoryId, '');
-        } else {
-          query = query.concat(',', categoryId);
-        }
-      } else {
-        query = categoryId;
-      }
-    } else if (query) {
-      if (!target.checked) {
-        query = query.replace(target.name, '');
-      } else {
-        query = query.concat(',', target.name);
-      }
-    } else {
-      query = target.name;
-    }
-    if (query) {
-      searchParams.set(queryName, query);
-    } else {
-      searchParams.delete(queryName);
-    }
-    searchParams.set(page, 1);
-    history.push(`?${searchParams.toString()}`);
-  };
-
-  const handleFilterClear = (setFilter, queryName) => {
-    searchParams.set(page, 1);
-    dispatch(setFilter([]));
-    searchParams.delete(queryName);
-    history.push(`?${searchParams.toString()}`);
-  };
-  const filtersOptions = {
-    categories: {
-      filterName: CATERGORY_TEXT[language].value,
-      productFilter: categoryFilter,
-      list: map(
-        filterData.categories,
-        (category) => category.name[language].value
-      ),
-      categories: filterData.categories,
-      filterAction: setCategoryFilter,
-      labels: URL_QUERIES_NAME.categoryFilter,
-      clearFilter: () =>
-        handleFilterClear(setCategoryFilter, URL_QUERIES_NAME.categoryFilter),
-      filterHandler: (e) =>
-        handleFilterChange(
-          e,
-          URL_QUERIES_NAME.categoryFilter,
-          filterData.categories
-        )
-    },
-    models: {
-      filterName: MODEL_TEXT[language].value,
-      productFilter: modelsFilter,
-      list: map(filterData.models, (model) => model.name[language].value),
-      categories: filterData.models,
-      filterAction: setModelsFilter,
-      labels: URL_QUERIES_NAME.modelsFilter,
-      clearFilter: () =>
-        handleFilterClear(setModelsFilter, URL_QUERIES_NAME.modelsFilter),
-      filterHandler: (e) =>
-        handleFilterChange(e, URL_QUERIES_NAME.modelsFilter, filterData.models)
-    },
-    patterns: {
-      filterName: PATTERN_TEXT[language].value,
-      productFilter: patternsFilter,
-      list: map(filterData.patterns, (pattern) => pattern.name[language].value),
-      categories: filterData.patterns,
-      filterAction: setPatternsFilter,
-      labels: URL_QUERIES_NAME.patternsFilter,
-      clearFilter: () =>
-        handleFilterClear(setPatternsFilter, URL_QUERIES_NAME.patternsFilter),
-      filterHandler: (e) =>
-        handleFilterChange(
-          e,
-          URL_QUERIES_NAME.patternsFilter,
-          filterData.patterns
-        )
-    }
-  };
   const handleClearFilter = () => {
     const sortQuery = searchParams.get(sort);
     const quantityPerPage = searchParams.get(countPerPage);
